refactor bitcoin::builders
